feat: 集成 Skill 工具、修复检索与数据库字符集问题 NormalChat 集成 Skill 能力 新增 skill_tool.go，基于 skill.NewLocalBackend 实现按需加载 SKILL.md 在 flow.go 中为 ReAct Agent 注入 skill 工具 在 config.yaml 中新增 skills.baseDir 配置 创建 backend/skills 目录，并将 NormalChat 备份为 NormalChat_backup 技能目录调整 移除 frontend-beautifier 技能 新增 high-eq-communication 技能（逻辑清晰、高情商话术） 检索 cron_id 报错修复 在 EsHit2Document 与 helpers.go 中处理 cron_id 字段 将未知字段的 default 分支改为 continue，避免 schema 扩展导致检索失败 数据库字符集修复 在 database.default 与 db.mysql 连接串中增加 charset=utf8mb4 新增 scripts/fix_charset_utf8mb4.sql 用于修复已有表字符集 403 错误提示优化 在 openai.go 中为 403/RequestError 增加更明确的错误提示 日志中增加模式（CoachChat/NormalChat）信息

+++ backend/internal/logic/cron/scheduler.go
@@ -68,7 +68,7 @@ func AddJob(ctx context.Context, schedule *entity.KnowledgeBaseCronSchedule) err
 }
 
 // RemoveJob 从调度器移除任务
-func RemoveJob(id int) {
+func RemoveJob(id int64) {
 	if val, ok := jobMap.Load(id); ok {
 		entryID := val.(cron.EntryID)
 		CronScheduler.Remove(entryID)
